+++ docs/src/pages/components/speed-dial/ControlledOpenSpeedDial.tsx
@@ -1,18 +1,14 @@
 import * as React from 'react';
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
+import { makeStyles, Theme } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import SpeedDial from '@material-ui/lab/SpeedDial';
 import SpeedDialIcon from '@material-ui/lab/SpeedDialIcon';
 import SpeedDialAction from '@material-ui/lab/SpeedDialAction';
 import FileCopyIcon from '@material-ui/icons/FileCopyOutlined';
 import SaveIcon from '@material-ui/icons/Save';
-import PrintIcon from '@material-ui/icons/Print';
-import ShareIcon from '@material-ui/icons/Share';
-import FavoriteIcon from '@material-ui/icons/Favorite';
 import EditIcon from '@material-ui/icons/Edit';
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
+const useStyles = makeStyles((theme: Theme) => ({
     root: {
       height: 380,
       transform: 'translateZ(0px)',
@@ -29,33 +25,43 @@ const useStyles = makeStyles((theme: Theme) =>
 const actions = [
   { icon: <FileCopyIcon />, name: 'Copy' },
   { icon: <SaveIcon />, name: 'Save' },
-  { icon: <PrintIcon />, name: 'Print' },
-  { icon: <ShareIcon />, name: 'Share' },
-  { icon: <FavoriteIcon />, name: 'Like' },
 ];
 
-export default function OpenIconSpeedDial() {
+export default function ControlledOpenSpeedDial() {
   const classes = useStyles();
+  const [open, setOpen] = React.useState(false);
   const [hidden, setHidden] = React.useState(false);
 
   const handleVisibility = () => {
     setHidden((prevHidden) => !prevHidden);
   };
 
+  const handleOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
   return (
     <div className={classes.root}>
       <Button onClick={handleVisibility}>Toggle Speed Dial</Button>
       <SpeedDial
-        ariaLabel="SpeedDial openIcon example"
+        ariaLabel="SpeedDial uncontrolled open example"
         className={classes.speedDial}
         hidden={hidden}
         icon={<SpeedDialIcon openIcon={<EditIcon />} />}
+        onClose={handleClose}
+        onOpen={handleOpen}
+        open={open}
       >
         {actions.map((action) => (
           <SpeedDialAction
             key={action.name}
             icon={action.icon}
             tooltipTitle={action.name}
+            onClick={handleClose}
           />
         ))}
       </SpeedDial>
