+++ docs/src/pages/components/speed-dial/SpeedDials.js
@@ -52,7 +52,6 @@ export default function SpeedDials() {
   const classes = useStyles();
   const [direction, setDirection] = React.useState('up');
 
-  const [open, setOpen] = React.useState(false);
   const [hidden, setHidden] = React.useState(false);
 
   const handleDirectionChange = (event) => {
@@ -61,14 +60,6 @@ export default function SpeedDials() {
 
   const handleHiddenChange = (event) => {
     setHidden(event.target.checked);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  const handleOpen = () => {
-    setOpen(true);
   };
 
   return (
@@ -104,9 +95,6 @@ export default function SpeedDials() {
           className={classes.speedDial}
           hidden={hidden}
           icon={<SpeedDialIcon />}
-          onClose={handleClose}
-          onOpen={handleOpen}
-          open={open}
           direction={direction}
         >
           {actions.map((action) => (
@@ -114,7 +102,6 @@ export default function SpeedDials() {
               key={action.name}
               icon={action.icon}
               tooltipTitle={action.name}
-              onClick={handleClose}
             />
           ))}
         </SpeedDial>
